Allow for LUKS-encryption via tpm2
This solves the attack vector, that someone is taking the disks out of the decomission bucket and would be able to read anything without the system. The machine itself still boots without any credendials to enter.

As soon as `configure-tpm-for-luks.service` is done with creating the dracut based initrd, the system is actually already in the requested state.

`clevis` is used as a nice way to get all the necessary programs and libraries for `sd-cryptsetup` inside the initrd, as ubuntu _does not_ ship the tpm-tss2 darcut module.

Also, this system should already be secure boot ready thanks to `sicherboot`, but I didn't verified this, yet.